requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
